Rev.0/driver/simulator/requirements.txt
pyserial-asyncio>=0.6
yapf>=0.43.0
aioconsole>=0.8.1